move top-up components to UI folder
