+++ universal-login-react/src/ui/TopUp/TopUp.tsx
@@ -1,9 +1,9 @@
 import React, {useState} from 'react';
 import {OnGasParametersChanged, stringToEther} from '@universal-login/commons';
 import UniversalLoginSDK from '@universal-login/sdk';
-import {Safello} from '../../integration/Safello';
-import {Ramp} from '../../integration/Ramp';
-import {Wyre} from '../../integration/Wyre';
+import {Safello} from './OnRamp/Safello';
+import {Ramp} from './OnRamp/Ramp';
+import {Wyre} from './OnRamp/Wyre';
 import {TopUpComponentType} from '../../core/models/TopUpComponentType';
 import {ChooseTopUpMethod} from './ChooseTopUpMethod';
 import {ModalWrapper} from '../Modals/ModalWrapper';
