+++ universal-login-react/test/ui/Safello.tsx
@@ -2,7 +2,7 @@ import {expect} from 'chai';
 import React from 'react';
 import {mount} from 'enzyme';
 import {TEST_ACCOUNT_ADDRESS} from '@universal-login/commons';
-import {Safello} from '../../src/integration/Safello';
+import {Safello} from '../../src/ui/TopUp/OnRamp/Safello';
 
 describe('Safello', () => {
   it('should be mounted', async () => {
